SOC-1241 | Abstracting applyFilters logic to mixin and component
