+++ front/main/app/routes/discussion/forum.js
@@ -3,11 +3,14 @@ import DiscussionRouteUpvoteMixin from '../../mixins/discussion-route-upvote';
 import DiscussionForumModel from '../../models/discussion-forum';
 import DiscussionLayoutMixin from '../../mixins/discussion-layout';
 import DiscussionModerationRouteMixin from '../../mixins/discussion-moderation-route';
+import DiscussionFilteringRouteMixin from '../../mixins/discussion-filtering-route';
 
 export default DiscussionBaseRoute.extend(
 	DiscussionLayoutMixin,
 	DiscussionRouteUpvoteMixin,
-	DiscussionModerationRouteMixin, {
+	DiscussionModerationRouteMixin,
+	DiscussionFilteringRouteMixin,
+	{
 		discussionSort: Ember.inject.service(),
 		discussionEditor: Ember.inject.service(),
 
@@ -65,26 +68,6 @@ export default DiscussionBaseRoute.extend(
 			 */
 			setSortBy(sortBy) {
 				this.setSortBy(sortBy);
-			},
-
-			applyFilters(sortBy, onlyReported) {
-				const discussionSort = this.get('discussionSort'),
-					currentSortBy = discussionSort.get('sortBy');
-
-				let targetRoute;
-
-				if (sortBy !== currentSortBy) {
-					discussionSort.setSortBy(sortBy, onlyReported);
-					targetRoute = 'discussion.forum';
-				}
-
-				if (onlyReported === true) {
-					targetRoute = 'discussion.reported-posts';
-				}
-
-				if (targetRoute) {
-					return this.transitionTo(targetRoute, Mercury.wiki.id, sortBy);
-				}
 			},
 		}
 	}
