+++ front/main/app/routes/discussion/reported-posts.js
@@ -3,11 +3,14 @@ import DiscussionRouteUpvoteMixin from '../../mixins/discussion-route-upvote';
 import DiscussionReportedPostsModel from '../../models/discussion-reported-posts';
 import DiscussionLayoutMixin from '../../mixins/discussion-layout';
 import DiscussionModerationRouteMixin from '../../mixins/discussion-moderation-route';
+import DiscussionFilteringRouteMixin from '../../mixins/discussion-filtering-route';
+
 
 export default DiscussionBaseRoute.extend(
 	DiscussionLayoutMixin,
 	DiscussionRouteUpvoteMixin,
 	DiscussionModerationRouteMixin,
+	DiscussionFilteringRouteMixin,
 	{
 		discussionEditor: Ember.inject.service(),
 		discussionSort: Ember.inject.service(),
@@ -69,24 +72,6 @@ export default DiscussionBaseRoute.extend(
 			 */
 			setSortBy(sortBy) {
 				this.setSortBy(sortBy);
-			},
-
-			applyFilters(sortBy, onlyReported) {
-				const discussionSort = this.get('discussionSort'),
-					currentSortBy = discussionSort.get('sortBy');
-
-				let targetRoute;
-
-				if (sortBy !== currentSortBy) {
-					discussionSort.setSortBy(sortBy);
-					targetRoute = 'discussion.reported-posts';
-				}
-
-				if (onlyReported === false) {
-					targetRoute = 'discussion.forum';
-				}
-
-				return this.transitionTo(targetRoute, Mercury.wiki.id, sortBy);
 			},
 		}
 	}
